+++ src/Symfony/Component/HttpKernel/Controller/ArgumentResolverInterface.php
@@ -24,7 +24,9 @@ interface ArgumentResolverInterface
     /**
      * Returns the arguments to pass to the controller.
      *
+     * @param \ReflectionFunctionAbstract|null $reflector
+     *
      * @throws \RuntimeException When no value could be provided for a required argument
      */
-    public function getArguments(Request $request, callable $controller): array;
+    public function getArguments(Request $request, callable $controller/* , \ReflectionFunctionAbstract $reflector = null */): array;
 }