+++ src/Symfony/Component/HttpKernel/ControllerMetadata/ArgumentMetadataFactoryInterface.php
@@ -19,7 +19,9 @@ namespace Symfony\Component\HttpKernel\ControllerMetadata;
 interface ArgumentMetadataFactoryInterface
 {
     /**
+     * @param \ReflectionFunctionAbstract|null $reflector
+     *
      * @return ArgumentMetadata[]
      */
-    public function createArgumentMetadata(string|object|array $controller): array;
+    public function createArgumentMetadata(string|object|array $controller/* , \ReflectionFunctionAbstract $reflector = null */): array;
 }